Added Geometry class, cleanup & new methods.
Added new GeoGeometry class and merged the GeoDistance class into it. Added several new methods to GeoHashUtils to work directly with geo hash bit sets and cleaned up the apache code further. Removed the getPrecision method because it was too aggressive; we want the accurate center of a geohash and not a rough approximation of something that is in the geoHash.